fix(startup): defer watchdog until first heartbeat
